+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.1
+
+- Count assertions added
+- Custom dots space builder added
+- Dot animations added
+
 ## 1.0.0
 
 - First Release supporting `ScrollController`.